+++ backend/src/main/java/hanglog/city/domain/repository/CityRepository.java
@@ -1,6 +1,6 @@
-package hanglog.trip.domain.repository;
+package hanglog.city.domain.repository;
 
-import hanglog.trip.domain.City;
+import hanglog.city.domain.City;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CityRepository extends JpaRepository<City, Long> {
