+++ backend/src/main/java/hanglog/city/dto/response/CityResponse.java
@@ -2,7 +2,7 @@ package hanglog.city.dto.response;
 
 import static lombok.AccessLevel.PRIVATE;
 
-import hanglog.trip.domain.City;
+import hanglog.city.domain.City;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
