+++ backend/src/main/java/hanglog/city/dto/response/CityWithPositionResponse.java
@@ -1,8 +1,8 @@
-package hanglog.trip.dto.response;
+package hanglog.city.dto.response;
 
 import static lombok.AccessLevel.PRIVATE;
 
-import hanglog.trip.domain.City;
+import hanglog.city.domain.City;
 import java.math.BigDecimal;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
