+++ backend/src/main/java/hanglog/trip/dto/response/TripDetailResponse.java
@@ -3,8 +3,9 @@ package hanglog.trip.dto.response;
 import static hanglog.image.util.ImageUrlConverter.convertNameToUrl;
 import static lombok.AccessLevel.PRIVATE;
 
+import hanglog.city.dto.response.CityWithPositionResponse;
 import hanglog.share.domain.SharedTrip;
-import hanglog.trip.domain.City;
+import hanglog.city.domain.City;
 import hanglog.trip.domain.Trip;
 import java.time.LocalDate;
 import java.util.List;
