+++ backend/src/test/java/hanglog/share/fixture/ShareFixture.java
@@ -5,7 +5,7 @@ import static hanglog.share.domain.type.SharedStatusType.SHARED;
 import static hanglog.share.domain.type.SharedStatusType.UNSHARED;
 
 import hanglog.share.domain.SharedTrip;
-import hanglog.trip.domain.City;
+import hanglog.city.domain.City;
 import hanglog.trip.domain.DayLog;
 import hanglog.trip.domain.Trip;
 import java.math.BigDecimal;
